refactor(store): cart state is app state, not a feature, 6 -> 4 [GH-000]
store's two cross-feature imports are gone, and the way they went is
worth recording because my suggested fix was wrong.

The check script proposed inverting one direction with a callback prop.
That does not work. ProductGrid renders the cards and lives in
features/products too, so passing the cart state down from there moves
the import up a level and leaves it a violation. It only helps if the
threading goes all the way to a composition root, which means every
card's cart state travelling through props -- worse code bought with a
boundary count.

What was actually wrong is where the cart's state was filed. CartProvider
and FlyToCartProvider are mounted in app/[locale]/providers.tsx, app-wide,
and useAddToCart is needed on every product card. A context mounted at the
app root and used by two features is application state, not one feature's
business. So it moved:

  shared/application/cart/   CartContext, FlyToCartContext, cartReducer,
                             cartCookiePersistence, cartValidation,
                             useAddToCart, useFlyToCart
  shared/domain/cart.ts      CartItem and CartState, which both the state
                             and the drawer need
  shared/infrastructure/     fetchStoreProductsByIds -- the cart hydrates
                             cookie ids into records, so the query is the
                             interface between two features rather than
                             the catalogue's private business

features/cart is now the drawer that reads them, which is what it always
was. Moved with git mv so the history follows the files. 21 files rewired.

Four remain, all payments: assigned-orders and received-orders import
each other, and separating them means lifting the order-management domain
into shared rather than moving an import. Still a design decision.

lint 0, typecheck 12/12, store 350 passing, full suite green.
