+++ apps/store/src/features/cart/index.ts
@@ -1,10 +1,6 @@
-export { CartProvider, useCart } from "./application/CartContext";
-export {
-  FlyToCartProvider,
-  useFlyToCartContext,
-} from "./application/FlyToCartContext";
-export { useAddToCart } from "./application/hooks/useAddToCart";
-export type { CartItem, CartState } from "./domain/types";
-
-// Presentation
+// The cart drawer. The cart's *state* is not here: CartProvider, useCart and
+// useAddToCart live in shared/application/cart, because the providers are
+// mounted app-wide in app/[locale]/providers.tsx and features/products needs
+// useAddToCart on every product card. A context two features share, mounted at
+// the app root, is application state rather than one feature's business.
 export { CartDrawer } from "./presentation/components/CartDrawer";
