+++ apps/store/src/features/cart/presentation/components/CartDrawer.tsx
@@ -8,10 +8,10 @@ import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetTrigger } from "ui";
 
 import { CartItemRow } from "./CartItemRow";
 
-import { useCart } from "@/features/cart/application/CartContext";
-import { useFlyToCartContext } from "@/features/cart/application/FlyToCartContext";
 import { groupCartBySeller } from "@/features/cart/application/groupBySeller";
 import { useSellerProfiles } from "@/features/cart/application/hooks/useSellerProfiles";
+import { useCart } from "@/shared/application/cart/CartContext";
+import { useFlyToCartContext } from "@/shared/application/cart/FlyToCartContext";
 import { appUrls } from "@/shared/infrastructure/config";
 
 const BADGE_OVERFLOW_THRESHOLD = 99;
