+++ apps/store/src/features/products/presentation/components/ProductDetail/HeroSection.tsx
@@ -8,11 +8,11 @@ import { ImageGallery } from "./ImageGallery";
 import { PriceBlock } from "./PriceBlock";
 import { RatingStars } from "./RatingStars";
 
-import { useAddToCart } from "@/features/cart";
 import {
   isProductAvailable,
   type Product,
 } from "@/features/products/domain/types";
+import { useAddToCart } from "@/shared/application/cart";
 import type { CategoryTheme } from "@/shared/domain/categoryConstants";
 
 interface HeroSectionProps {
