+++ apps/store/src/shared/application/cart/FlyToCartContext.tsx
@@ -2,7 +2,7 @@
 
 import { createContext, useContext } from "react";
 
-import { useFlyToCart } from "./hooks/useFlyToCart";
+import { useFlyToCart } from "./useFlyToCart";
 
 type FlyToCartContextValue = ReturnType<typeof useFlyToCart>;
